Refactor toolkit kernel module: rename files and functions & decouple files  responsabilities
