+++ packages/toolkit/test/dao/kernel/appProxy.test.js
@@ -1,6 +1,6 @@
 import test from 'ava'
 //
-import { getAppProxyAddressFromReceipt } from '../../../src/dao/kernel/kernel'
+import { getAppProxyAddressFromReceipt } from '../../../src/dao/kernel/appProxy'
 
 test('getAppProxyAddressFromReceipt', t => {
   /**
